Master

+++ Netch/Utils/ModeHelper.cs
@@ -1,12 +1,14 @@
 using System;
 using System.IO;
 using System.Linq;
+using System.Reactive.Linq;
 using Netch.Controllers;
 using Netch.Enums;
 using Netch.Interfaces;
 using Netch.Models;
+using Netch.Servers;
 using Netch.Servers.Shadowsocks;
-using Netch.Servers.Socks5;
+using Serilog;
 
 namespace Netch.Utils
 {
@@ -14,32 +16,47 @@ namespace Netch.Utils
     {
         public const string DisableModeDirectoryFileName = "disabled";
 
+        private static FileSystemWatcher _fileSystemWatcher = null!;
+
         public static string ModeDirectoryFullName => Path.Combine(Global.NetchDir, "mode");
 
-        private static readonly FileSystemWatcher FileSystemWatcher;
-
-        public static bool SuspendWatcher { get; set; } = false;
-
-        static ModeHelper()
+        public static bool SuspendWatcher
         {
-            FileSystemWatcher = new FileSystemWatcher(ModeDirectoryFullName)
+            get => _fileSystemWatcher.EnableRaisingEvents;
+            set => _fileSystemWatcher.EnableRaisingEvents = value;
+        }
+
+        public static void InitWatcher()
+        {
+            _fileSystemWatcher = new FileSystemWatcher(ModeDirectoryFullName)
             {
-                NotifyFilter = NotifyFilters.LastWrite | NotifyFilters.Size | NotifyFilters.FileName,
+                NotifyFilter = NotifyFilters.LastWrite | NotifyFilters.FileName | NotifyFilters.DirectoryName,
                 IncludeSubdirectories = true,
                 EnableRaisingEvents = true
             };
 
-            FileSystemWatcher.Changed += OnModeChanged;
-            FileSystemWatcher.Created += OnModeChanged;
-            FileSystemWatcher.Deleted += OnModeChanged;
-            FileSystemWatcher.Renamed += OnModeChanged;
+            var created = Observable.FromEventPattern<FileSystemEventHandler, FileSystemEventArgs>(h => _fileSystemWatcher.Created += h,
+                    h => _fileSystemWatcher.Created -= h)
+                .Select(x => x.EventArgs);
+
+            var changed = Observable.FromEventPattern<FileSystemEventHandler, FileSystemEventArgs>(h => _fileSystemWatcher.Changed += h,
+                    h => _fileSystemWatcher.Changed -= h)
+                .Select(x => x.EventArgs);
+
+            var deleted = Observable.FromEventPattern<FileSystemEventHandler, FileSystemEventArgs>(h => _fileSystemWatcher.Deleted += h,
+                    h => _fileSystemWatcher.Deleted -= h)
+                .Select(x => x.EventArgs);
+
+            var renamed = Observable.FromEventPattern<RenamedEventHandler, RenamedEventArgs>(h => _fileSystemWatcher.Renamed += h,
+                    h => _fileSystemWatcher.Renamed -= h)
+                .Select(x => x.EventArgs);
+
+            var o = Observable.Merge(created, deleted, renamed, changed);
+            o.Throttle(TimeSpan.FromSeconds(3)).Subscribe(_ => OnModeChange(), exception => Log.Error(exception, "FileSystemWatcherError"));
         }
 
-        private static void OnModeChanged(object sender, FileSystemEventArgs e)
+        private static void OnModeChange()
         {
-            if (SuspendWatcher)
-                return;
-
             Load();
             Global.MainForm.LoadModes();
         }
@@ -58,23 +75,19 @@ namespace Netch.Utils
             return Path.Combine(ModeDirectoryFullName, relativeName);
         }
 
-        /// <summary>
-        ///     从模式文件夹读取模式
-        /// </summary>
         public static void Load()
         {
             Global.Modes.Clear();
-            LoadModeDirectory(ModeDirectoryFullName);
-
+            LoadCore(ModeDirectoryFullName);
             Sort();
         }
 
-        private static void LoadModeDirectory(string modeDirectory)
+        private static void LoadCore(string modeDirectory)
         {
             try
             {
                 foreach (var directory in Directory.GetDirectories(modeDirectory))
-                    LoadModeDirectory(directory);
+                    LoadCore(directory);
 
                 // skip Directory with a disabled file in
                 if (File.Exists(Path.Combine(modeDirectory, DisableModeDirectoryFileName)))
@@ -87,7 +100,7 @@ namespace Netch.Utils
                     }
                     catch (Exception e)
                     {
-                        Global.Logger.Warning($"Load mode \"{file}\" failed: {e.Message}");
+                        Log.Warning(e, "Load mode \"{FileName}\" failed", file);
                     }
             }
             catch
@@ -106,7 +119,11 @@ namespace Netch.Utils
             if (mode.FullName == null)
                 throw new ArgumentException(nameof(mode.FullName));
 
-            File.Delete(mode.FullName);
+            Global.MainForm.ModeComboBox.Items.Remove(mode);
+            Global.Modes.Remove(mode);
+
+            if (File.Exists(mode.FullName))
+                File.Delete(mode.FullName);
         }
 
         public static bool SkipServerController(Server server, Mode mode)
@@ -142,7 +159,7 @@ namespace Netch.Utils
                 case ModeType.Pcap2Socks:
                     return new PcapController();
                 default:
-                    Global.Logger.Error("未知模式类型");
+                    Log.Error("未知模式类型");
                     throw new MessageException("未知模式类型");
             }
         }
